create modules: prior, likelihood, posterior
