+++ octopus/__init__.py
@@ -1,1 +1,4 @@
 from .core import *
+from .prior import *
+from .likelihood import *
+from .posterior import *
